requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
pydantic_geojson
pydantic_partial
python-dotenv
python-multipart
supabase
prometheus-client
rasterio
rio-cogeo
fire
pillow
